IGVF-3449-fix-report-links
* Convert input-file-sets.js to Typescript.
* Use different report links depending on the file-set type.
* Make `inputForQuery` property required, as I always have it anyway.
* Change Jest tests using the SampleObject type to have the now-required `summary` property.
