+++ lib/__tests__/data-use-limitation.test.ts
@@ -22,6 +22,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         institutional_certificates: [],
         release_timestamp: "2013-07-19T14:05:32.451982+00:00",
         status: "released",
+        summary: "A sample of primary cells.",
         uuid: "f8aa2bdd-e30a-4f22-893c-49c21f86824c",
         virtual: false,
       },
@@ -33,6 +34,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         institutional_certificates: [],
         release_timestamp: "2007-11-03T06:22:48.190374+00:00",
         status: "released",
+        summary: "Another sample of primary cells.",
         uuid: "42245db0-2e20-491d-a263-900db67dd152",
         virtual: false,
       },
@@ -51,6 +53,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         institutional_certificates: [],
         release_timestamp: "2013-07-19T14:05:32.451982+00:00",
         status: "released",
+        summary: "A sample of primary cells.",
         uuid: "4fce0ecd-b14c-4dab-b4b2-a5c8bea74d0f",
         virtual: false,
       },
@@ -77,6 +80,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2011-03-15T22:11:06.087000+00:00",
         status: "released",
+        summary: "Another sample of primary cells.",
         uuid: "cf77059d-ea0c-4e4d-b708-cb04877cc95d",
         virtual: false,
       },
@@ -110,6 +114,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2013-07-19T14:05:32.451982+00:00",
         status: "released",
+        summary: "A sample of primary cells.",
         uuid: "4fce0ecd-b14c-4dab-b4b2-a5c8bea74d0f",
         virtual: false,
       },
@@ -151,6 +156,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2011-03-15T22:11:06.087000+00:00",
         status: "released",
+        summary: "Another sample of primary cells.",
         uuid: "cf77059d-ea0c-4e4d-b708-cb04877cc95d",
         virtual: false,
       },
@@ -185,6 +191,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2013-07-19T14:05:32.451982+00:00",
         status: "released",
+        summary: "A sample of primary cells.",
         uuid: "4fce0ecd-b14c-4dab-b4b2-a5c8bea74d0f",
         virtual: false,
       },
@@ -226,6 +233,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2011-03-15T22:11:06.087000+00:00",
         status: "released",
+        summary: "Another sample of primary cells.",
         uuid: "cf77059d-ea0c-4e4d-b708-cb04877cc95d",
         virtual: false,
       },
@@ -243,6 +251,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         accession: "IGVFSM0000SAMP",
         release_timestamp: "2013-07-19T14:05:32.451982+00:00",
         status: "released",
+        summary: "A sample of primary cells.",
         uuid: "4fce0ecd-b14c-4dab-b4b2-a5c8bea74d0f",
         virtual: false,
       },
@@ -282,6 +291,7 @@ describe("Test collectDataUseLimitationSummariesFromSamples", () => {
         ],
         release_timestamp: "2011-03-15T22:11:06.087000+00:00",
         status: "released",
+        summary: "Another sample of primary cells.",
         uuid: "cf77059d-ea0c-4e4d-b708-cb04877cc95d",
         virtual: false,
       },
